requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.0+cu128
torchvision==0.22.0+cu128
torchaudio==2.7.0+cu128
accelerate>=0.34.0
datasets>=2.19.0
peft>=0.13.0
transformers>=4.45.0
tqdm>=4.66.0
pyyaml>=6.0.0
bitsandbytes>=0.44.0; platform_system != "Windows"
